src: humanize third hundred-region batch [1,345 of 1,999]

+++ src/08019ebc.c
@@ -3,7 +3,8 @@
 extern u8 Data_0802a2e0[];
 extern u8 Data_08029ee4[];
 
+/* 2つの境界アドレス間を4バイト単位で数える。 */
 s32 Func_08019ebc(void)
 {
-    return (Data_0802a2e0 - Data_08029ee4) >> 2;
+    return (s32)((u32)Data_0802a2e0 - (u32)Data_08029ee4) >> 2;
 }
